OAK-440 - Finished implementing a fix
The problem was every concurrent update was assumed to be a conflict
byy default. This caused CommitCommand's number of retries to run out
quickly. The fix was to check for affected paths in concurrent updates
and internally retry again (without using number of retries) when there
is no conflict. I tried to find a way to not even retry (by adjusting
already persisted commits and their baseRevIds) but that got really
complicated and not even sure if it's a proper fix, so trying
internally seemed like a safer solution.