OAK-4412 - Lucene hybrid index
Instead of holding the reference to previous reader upon creation obtain it every time. This is required as previous NRTIndex might remain in use while new NRTIndex is getting created (IndexTracker would first open new IndexNode and then close older one). So to ensure we do not miss out on any indexed content refetch the old index everytime.

if the previous index is not written to this should be cheap as same reader would be returned

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1760880 13f79535-47bb-0310-9956-ffa450edef68