OAK-7066: Active deletion blob list files can grow too large due to inlined blobs
While the issue is applicable to blob store. But, we can't cheaply distinguish between inlined and non-inlined data for BlobStore.
So, we're just fixing the issue for DataStore.
For BlobStore, all ids would still get marked as deletd - that's asserted in OakDirectoryTestBase#markAllBlobsFromBlobStoreAsDeleted
(that test should likely be removed updated if we fix it for blob store too later)

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1818545 13f79535-47bb-0310-9956-ffa450edef68