OAK-7284: Reindexing using --doc-traversal-mode can hit ConcurrentModificationException during aggregation
CursorableLinkedList led to heavy drop in performance as we weren't closing cursor.
And without that, the class would broadcast all changes to all open cursors.

Adding lifecycle mgmt was a mess, so cooking up our own poor man link list for our own purpose.

Btw, the list impl is not generics - because YAGNI

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1825523 13f79535-47bb-0310-9956-ffa450edef68